Types and Functions

Index of all types and functions

SellmeierFit.SellmeierModel

SellmeierFit.SellmeierModel

SellmeierFit.SpectralVariable

Base.length

SellmeierFit.fit_sellmeier

SellmeierFit.fit_sellmeier

SellmeierFit.fit_sellmeier

SellmeierFit.read

Types

SellmeierFit.SellmeierModel — Type

SellmeierModel{N}

Sellmeier model representing the Sellmeier equation ε(λ) = 1 + ∑ᵢ [Bᵢ λ² / (λ² - Cᵢ)], where the sum is for 1 ≤ i ≤ N.

Fields

str::SVector{N,Float64}: str[i] is the strength of the ith term in the Sellmeier equation; corresponds to Bᵢ.

λres::SVector{N,Float64}: λres[i] is the resonance wavelength of the ith term in the Sellmeier equation; corresponds to √Cᵢ.

SellmeierFit.SellmeierModel — Method

(::SellmeierModel)(λ::Real)

Evaluate the Sellmeier equation represented by the given SellmeierModel at a wavelngth λ in units of meters.

SellmeierFit.SpectralVariable — Type

SpectralVariable

Used to specify the type of the spectral variable which the refractive index is a function of. The available instances are WAVELENGTH for wavelength, FREQUENCY for frequency, and ENERGY for photon energy.

Functions

Base.length — Method

Base.length(sm::SellmeierModel)

Return the number of terms in the sum of the Sellmeier equation represented by the given SellmeierModel.

SellmeierFit.fit_sellmeier — Method

fit_sellmeier(λ::AbsVecFloat, ε::AbsVecFloat, Nₙ::Integer, Nₚ::Integer)

Fit the dielectric constant data ε sampled at wavelengths λ to the Sellmeier equation with exactly Nₙ terms below and Nₚ terms above the range of λ.

SellmeierFit.fit_sellmeier — Method

fit_sellmeier(λ::AbsVecFloat, ε::AbsVecFloat; Nmax::Integer=10, rtol_λres::Real=1e-3)

Fit the dielectric constant data ε sampled at wavelengths λ to the Sellmeier equation with up to Nmax terms. If λres of different terms have relative difference less than rtol_λres, they are considered the same term and merged.

SellmeierFit.fit_sellmeier — Method

fit_sellmeier(λ::AbsVecFloat, ε::AbsVecFloat, Val(N))

Fit the dielectric constant data ε sampled at wavelengths λ to the Sellmeier equation with exactly N terms. Out of the N terms, some are below and the other are above the range of λ.

SellmeierFit.read — Method

read(path::String; <keyword arguments>)

Read the refractive index file at path, and return the wavelength vector λ and the dielectric constant vector ε as a named tuple.

Keywords

s_var::SpectralVariable=WAVELENGTH: WAVELENGTH, FREQUENCY, or ENERGY if the refractive index file describes refractive index as a function of wavelength, frequency, or photon energy.

unit_prefix::Real=micro: the prefix to multiply to the values of the spectral variable to convert their units to the SI units. For example, 1e-6 for wavelengths in units of µm; 1.602176643 for photon energies in units of eV.

s_col::Integer=1: the index of the spectral variable column in the refractive index file.

n_col::Integer=2: the index of the refractive index column in the refractive index file.

kwargs...: the keyword arguments to pass to CSV.read()